refactor(core): Create VenokContext for all executions (delete HTTP)
